requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
inflection==0.5.1
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
PyJWT==2.7.0
PySocks==1.7.1
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
uritemplate==4.1.1
urllib3==2.0.4
wsproto==1.2.0
